pyproject.toml
[tool.poetry]
name = "quintet-poker"
version = "0.1.0"
description = "A simple poker game using FastAPI and SQLite"
authors = ["Daniel(Donghyun) Sung <donghyun.main@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.0"
uvicorn = "^0.15.0"
sqlalchemy = "^1.4.23"
authlib = "^1.3.2"
httpx = "^0.27.2"
itsdangerous = "^2.2.0"
python-dotenv = "^1.0.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = "^21.7b0"
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
